added project config
