+++ main.go
@@ -5,20 +5,18 @@ import (
 	"encoding/hex"
 	"fmt"
 	"github.com/go-playground/webhooks/v6/github"
-	"gopkg.in/yaml.v3"
 	"log"
 	"log/slog"
 	"net/http"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 )
 
 const (
-	WEBHOOK_PATH = "/webhooks"
-	CACHE_DIR    = "/tmp"
-	LOG_DIR      = "/log"
+	WEBHOOK_PATH = "~/webhooks"
+	CACHE_DIR    = "~/tmp"
+	LOG_DIR      = "~/log"
 )
 
 type Context struct {
@@ -29,10 +27,6 @@ type Context struct {
 	branchSha     string
 	commitSha     string
 	repoBranchSha string
-}
-
-type ProjectConf struct {
-	Phase map[string]string `yaml:"phase"`
 }
 
 func generateContext(repo, ref, defaultBranch, commitSha string) Context {
@@ -120,13 +114,6 @@ func handleDown(ctx Context) error {
 	return nil
 }
 
-func execCmd(command string, args ...string) error {
-	cmd := exec.Command(command, args...)
-	cmd.Stdout = log.Writer()
-	cmd.Stderr = log.Writer()
-	return cmd.Run()
-}
-
 func main() {
 	err := os.MkdirAll(LOG_DIR, 0755)
 	if err != nil {
@@ -194,55 +181,4 @@ func main() {
 		fmt.Fprintln(w, "")
 	})
 	http.ListenAndServe(":80", nil)
-}
-
-func (c *ProjectConf) getPhaseEnv(ctx Context) string {
-	// get value of branch in c.Phase if not exist use "*"
-	environment := c.Phase[ctx.branch]
-	if environment == "" {
-		environment = c.Phase["*"]
-	}
-	return environment
-}
-
-func getProjectConf(ctx Context) (projectConf ProjectConf, err error) {
-	c := ProjectConf{}
-
-	cacheDir := filepath.Join(CACHE_DIR, ctx.commitSha+ctx.defaultBranch)
-	defer func() {
-		err := os.RemoveAll(cacheDir)
-		if err != nil {
-			log.Printf("failed to remove `%s`\n", cacheDir)
-		}
-	}()
-
-	err = execCmd("git", "clone", "--branch", ctx.defaultBranch, "--single-branch", ctx.repo, cacheDir)
-	if err != nil {
-		return c, fmt.Errorf("failed `git clone`: %w", err)
-	}
-
-	envPklFilePath := filepath.Join(cacheDir, "env.pkl")
-	ymlPklFilePath := filepath.Join(cacheDir, "env.yml")
-
-	err = execCmd("stat", envPklFilePath)
-	if err != nil {
-		return c, nil
-	}
-
-	err = execCmd("pkl", "eval", envPklFilePath, "--format", "yaml", "--output-path", ymlPklFilePath, "--property", "branch="+ctx.branch)
-
-	if err != nil {
-		return c, fmt.Errorf("failed `pkl eval`: %w", err)
-	}
-	out, err := os.ReadFile(ymlPklFilePath)
-	if err != nil {
-		return c, fmt.Errorf("failed to read %+v", err)
-	}
-
-	err = yaml.Unmarshal(out, &c)
-	if err != nil {
-		return c, fmt.Errorf("failed to unmarshal %+v", err)
-	}
-
-	return c, nil
 }
