+++ util.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"log"
+	"os/exec"
+)
+
+func execCmd(command string, args ...string) error {
+	cmd := exec.Command(command, args...)
+	cmd.Stdout = log.Writer()
+	cmd.Stderr = log.Writer()
+	return cmd.Run()
+}
